Remove CloudPrint and align list
CloudPrint service ends with the end of the year and the install option has been disabled in DietPi to not offer an option that is functional for some ~2 weeks only.

WiFi HotSpot, Tor HotSpot and No-IP have been moved to the Advanced Networking category. Move it to the top of the list to stay with alphabetical order. WiFi HotSpot/AP category has been removed.