wordlist cleanup
Add names with capital first letter, so it is not allowed to be used with non-capital first letter.

Add non-names with non-capital first letters, so it is allowed to use them with capital (beginning of a sentence) AND non-capital first letter.

Add abbreviations in full capital letters only. Whenever they are used as part of a file name or extension, or console command or output, where they appear in lower-case letters they shall be quoted as inline code.