Minor SBC install clarifications
When using Rufus, the example instructions from the UEFI tab must not be followed. In case of SBC MBR full drive images, all options are greyed out correctly and after selecting image and target device, hitting START is all that is possible and required.

SBC images do not necessarily need to be flashed to an SD card, but a USB drive works as well. Use "device" or "drive resp. SD card" and similar to make that clear.

On first boot, not only the root partition and filesystem is expanded, but also other initial setup steps are done, like applying locale settings from dietpi.txt, re-creating fresh SSH host keys and system UUID and such. Mention "initial setup steps" to cover that.