Fix local link base URL
The base URL is not suffixed with the file path, hence it is wrong as fast as sub directories are processed. It is hence required to loop through all directories and check those individually with the directory as base path each.

Remove verbose flag. It is ignored when STDOUT and STDERR are redirected.

Do not serve redirects, so we can find outdated links.