Link Bullseye upgrade blog post
and additionally fix/change all user profile links from GitHub to phpBB, when the issue was reported in our forum. While some users might have a GitHub profile with identical name, this is of course not always the case so we are currently linking to some wrong or not existing profiles.