Remove misleading info about Bullseye upgrade
No one should ever get the impression that upgrading to Bullseye would be required for any DietPi update: Buster is supported at least until summer 2023 and even when the underlying Debian is unsupported, the migration to a dedicated support branch and related info will done smoothly. Generally DietPi updates and Debian upgrades are completely independent, and we will inform early when a Debian version is about to loose support, but the the DietPi release notes are not the right place for this (aside of the exact DietPi version which drops a certain Debian version, e.g. Debian Stretch with DietPi 8.0).