Update download links to Bullseye
And replace Odroid N1 with C4/HC4 since we're going to remove the dediated Odroid N1 hardware ID and image (not a single reported install, that SBC was never really released aside of a few developer samples) and C4/HC4 was simply missing.

Also update/simplify the PREP script usage instructions by removing everything about the alternative branches: All end users should use the master branch only, and those who are able to deal with the possible implications of using beta or dev branch for either PREP or the final system will know how to select it, respectively, respectively, we make this the condition now :).