Add meaningful title and description meta tags (#676)
By default the navigation menu text is used as page title, which is not always self-explaining, like the "Overview" pages for software and releases. As description meta tag, the global documentation description "Official documentation pages for DietPi OS" is used, so this adds more specific page descriptions.

Could or should be expanded to individual release notes and software categories, so this is a start to cover all top level pages. Title and description don't need to be given both, if the title (navigation menu text) is sufficiently self-explaining, the description alone can be given.

This also adds some internal links and mininal text changes.