Update VirtualHere USB drive warning
USB drives are not unmounted on service start, but when selecting them for use in VirtualHere clients, even if they are actively used with e.g. a swap file or database server. The USB device is shown like detached on the host system, processes with current I/O to the related filesystem will crash, leading to potential data loss. Even with a swap file on the drive, VirtualHere does not hesitate, leading to kernel errors and potential complete system or service crashes, depending on which data was swapped. Add info about how to ignore devices via config.