Remove all YouTube video embeddings (#781)
and replace them with simple links. This is due to the fact that even with youtube-nocookie, such embeddings violate GDPR as long as there is no interactive consent button before the embedding is shown. And this again would require custom JavaScript or another 3rd party + the direct embedding benefit gets lost for users when several clicks are required, compared to a single click on a YouTube link, if interested.

CI | Ignore Roon docs link checks: which return an invalid HTTP header, while it works well in browser