CI | Workaround for failing code fence + local IP exclude
Should be solved with next lychee release which adds automatic exclusion of nested verbatim tags. It's strange that this is currently inconsistently excluded, e.g. when downloading https://dietpi.com/docs/install/ to file, it is excluded, when checking online resource directly, it is not. And it was excluded some week(s) ago on our CI, and is not anymore, suddenly, without any related change to either our install.md/html and without any lychee release in the meantime... One explanation could be a Python module upgrade, leading to a slightly different HTML syntax, which however does not explain differnt local vs remote results.