Formatting beautification
Re-added line breaks.
mkdocs uses two spaces at the end of a line to insert a line break without starting a new paragraph (<br>).
A new paragraph is started with a blank line.

I re-added some simple line breaks we had before in the doc for a bit nicer formatting.