Update Snapcast AirPlay docs
Our Shairport Sync has STDOUT support builtin, hence no need to compile a new binary. The service does not need to be masked, but only disabled. Also make clear which config file to edit.

This was found when aiming to fix the Shairport Sync install URL, which does not exist anymore.