Update VNC server docs
Fix login credentials: We use VncAuth with default software password. Align info about how the server mode defines the needed screen number on the client, and the network port this relates to.

Fix commands used to change the VncAuth password.

We do not use RealVNC's vncserver-x11-serviced service anymore, hence that mode does not change how to check logs.

Remove the direct capture mode info: It is not mentioned in the RPi docs anymore, instead another page says it does not work with KMS enabled. We'd need to test before re-adding this info.

The docs require further restructure and alignment: TigerVNC and RealVNC behave pretty similar now. One can use `dietpi-autostart 16` in shared desktop mode to login as different user from there, and we should mention that `SOFTWARE_VNCSERVER_DISPLAY_INDEX` affects the screen index that needs to be appended to the IP/hostname on the client, as well as the TCP port. However, stopping after the most urgent fixes for now.